requirements.txt
python-telegram-bot
phidata
python-dotenv
